pyproject.toml
[project]
name = "bedrock-api-keys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.39.4",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
